Refactor another timeblock hook
